+++ WooCommerce/Classes/Tools/RequirementsChecker.swift
@@ -68,7 +68,7 @@ class RequirementsChecker {
     /// - parameter result: Closure param that is the result of the requirement check
     /// - parameter error: Closure param that is any error that occured while checking the WC version
     ///
-    static func checkMinimumWooVersion(for siteID: Int, onCompletion: ((_ result: RequirementCheckResult, _ error: Error?) -> Void)? = nil) {
+    static func checkMinimumWooVersion(for siteID: Int64, onCompletion: ((_ result: RequirementCheckResult, _ error: Error?) -> Void)? = nil) {
         let action = retrieveSiteAPIAction(siteID: siteID, onCompletion: onCompletion)
         ServiceLocator.stores.dispatch(action)
     }
@@ -90,7 +90,7 @@ private extension RequirementsChecker {
 
     /// Returns a `SettingAction.retrieveSiteAPI` action
     ///
-    static func retrieveSiteAPIAction(siteID: Int, onCompletion: ((RequirementCheckResult, Error?) -> Void)? = nil) -> SettingAction {
+    static func retrieveSiteAPIAction(siteID: Int64, onCompletion: ((RequirementCheckResult, Error?) -> Void)? = nil) -> SettingAction {
         return SettingAction.retrieveSiteAPI(siteID: siteID) { (siteAPI, error) in
             guard error == nil else {
                 DDLogError("⛔️ An error occurred while fetching API info for siteID \(siteID): \(String(describing: error))")
