+++ Yosemite/Yosemite/Actions/AvailabilityAction.swift
@@ -3,5 +3,5 @@
 public enum AvailabilityAction: Action {
     /// Checks if Stats v4 is available for the site.
     ///
-    case checkStatsV4Availability(siteID: Int, onCompletion: (_ isStatsV4Available: Bool) -> Void)
+    case checkStatsV4Availability(siteID: Int64, onCompletion: (_ isStatsV4Available: Bool) -> Void)
 }
